+++ src/main/scala/Day4.scala
@@ -10,7 +10,7 @@ object Day4 extends AoC:
       val mine = NumRe.findAllIn(chosen).toSet
       wins.intersect(mine).size
 
-  override def a(lines: Vector[String]): Int =
+  override def a(lines: Vector[String]): Long =
     val scores = for
       line <- lines
       ok    = wins(line)
@@ -20,7 +20,7 @@ object Day4 extends AoC:
     scores.sum
   end a
 
-  override def b(lines: Vector[String]): Int =
+  override def b(lines: Vector[String]): Long =
     val winMap             = lines.zipWithIndex.map(_.swap.rightMap(wins)).toMap
     def count(i: Int): Int = 1 + (i + 1 to i + winMap(i)).map(count).sum
     lines.indices.map(count).sum
